requirements.txt
wordcloud==1.5.0
numpy==1.16.0
scipy==1.1.0
tqdm==4.29.1
pandas==0.24.1
matplotlib==3.0.2
requests==2.21.0
langdetect==1.0.7
nltk==3.3
Pillow==5.4.1
beautifulsoup4==4.7.1
pyenchant==2.0.0
vaderSentiment==3.2.1
